refactored datetime widget
