+++ packages/material/test/renderers/MaterialDateTimeField.test.tsx
@@ -1,23 +1,17 @@
-import '../../../test/helpers/setup';
+import { initJsonFormsStore } from '@jsonforms/test';
 import * as React from 'react';
 import test from 'ava';
 import {
   ControlElement,
   getData,
   HorizontalLayout,
-  initJsonFormsStore,
   JsonSchema,
   update
 } from '@jsonforms/core';
 import HorizontalLayoutRenderer from '../../src/layouts/MaterialHorizontalLayout';
-import DateTimeField, { datetimeFieldTester } from '../../src/fields/material-datetime.field';
-import {
-  change,
-  findRenderedDOMElementWithTag,
-  renderIntoDocument,
-  scryRenderedDOMElementsWithTag
-} from '../../../test/helpers/binding';
+import DateTimeField, { datetimeFieldTester } from '../../src/fields/MaterialDateTimeField';
 import { Provider } from 'react-redux';
+import * as TestUtils from 'react-dom/test-utils';
 import * as moment from 'moment';
 
 test.beforeEach(t => {
@@ -33,9 +27,7 @@ test.beforeEach(t => {
   };
   t.context.uischema = {
     type: 'Control',
-    scope: {
-      $ref: '#/properties/foo',
-    },
+    scope: '#/properties/foo',
   };
 });
 test.failing('autofocus on first element', t => {
@@ -48,18 +40,14 @@ test.failing('autofocus on first element', t => {
     };
     const firstControlElement: ControlElement = {
         type: 'Control',
-        scope: {
-            $ref: '#/properties/firstDateTime'
-        },
+        scope: '#/properties/firstDateTime',
         options: {
             focus: true
         }
     };
     const secondControlElement: ControlElement = {
         type: 'Control',
-        scope: {
-            $ref: '#/properties/secondDateTime'
-        },
+        scope: '#/properties/secondDateTime',
         options: {
             focus: true
         }
@@ -80,12 +68,12 @@ test.failing('autofocus on first element', t => {
         schema,
         uischema
     });
-    const tree = renderIntoDocument(
+    const tree = TestUtils.renderIntoDocument(
         <Provider store={store}>
           <HorizontalLayoutRenderer schema={schema} uischema={uischema}/>
         </Provider>
     );
-    const inputs = scryRenderedDOMElementsWithTag(tree, 'input');
+    const inputs = TestUtils.scryRenderedDOMElementsWithTag(tree, 'input');
     t.not(document.activeElement, inputs[0]);
     t.is(document.activeElement, inputs[1]);
 });
@@ -93,9 +81,7 @@ test.failing('autofocus on first element', t => {
 test('autofocus active', t => {
     const uischema: ControlElement = {
         type: 'Control',
-        scope: {
-            $ref: '#/properties/foo'
-        },
+        scope: '#/properties/foo',
         options: {
             focus: true
         }
@@ -105,21 +91,19 @@ test('autofocus active', t => {
       schema: t.context.schema,
       uischema
     });
-    const tree = renderIntoDocument(
+    const tree = TestUtils.renderIntoDocument(
         <Provider store={store}>
           <DateTimeField schema={t.context.schema} uischema={uischema}/>
         </Provider>
     );
-    const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+    const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
     t.is(document.activeElement, input);
 });
 
 test('autofocus inactive', t => {
     const uischema: ControlElement = {
         type: 'Control',
-        scope: {
-            $ref: '#/properties/foo'
-        },
+        scope: '#/properties/foo',
         options: {
             focus: false
         }
@@ -129,33 +113,31 @@ test('autofocus inactive', t => {
       schema: t.context.schema,
       uischema
     });
-    const tree = renderIntoDocument(
+    const tree = TestUtils.renderIntoDocument(
         <Provider store={store}>
           <DateTimeField schema={t.context.schema} uischema={uischema}/>
         </Provider>
     );
-    const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+    const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
     t.false(input.autofocus);
 });
 
 test('autofocus inactive by default', t => {
     const uischema: ControlElement = {
         type: 'Control',
-        scope: {
-            $ref: '#/properties/foo'
-        }
+        scope: '#/properties/foo'
     };
     const store = initJsonFormsStore({
       data: t.context.data,
       schema: t.context.schema,
       uischema
     })
-    const tree = renderIntoDocument(
+    const tree = TestUtils.renderIntoDocument(
         <Provider store={store}>
           <DateTimeField schema={t.context.schema} uischema={uischema}/>
         </Provider>
     );
-    const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+    const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
     t.false(input.autofocus);
 });
 
@@ -224,13 +206,13 @@ test('render', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
 
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   t.is(input.type, 'text');
   t.is(input.value, '04.04.1980 13:37');
 });
@@ -241,14 +223,14 @@ test.cb('update via event', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   })
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   input.value = '12.04.1961 20:15';
-  change(input);
+  TestUtils.Simulate.change(input);
   setTimeout(
     () => {
       t.is(getData(store.getState()).foo, moment('1961-04-12 20:15').format());
@@ -264,12 +246,12 @@ test.cb('update via action', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   })
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   store.dispatch(update('foo', () => moment('1961-04-12 20:15').format()));
   setTimeout(
     () => {
@@ -286,12 +268,12 @@ test('update with null value', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   })
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   store.dispatch(update('foo', () => null));
   t.is(input.value, '');
 });
@@ -302,12 +284,12 @@ test('update with undefined value', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   })
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   store.dispatch(update('foo', () => undefined));
   t.is(input.value, '');
 });
@@ -318,12 +300,12 @@ test('update with wrong ref', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   })
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   store.dispatch(update('bar', () => 'Bar'));
   t.is(input.value, '04.04.1980 13:37');
 });
@@ -334,12 +316,12 @@ test('update with null ref', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   store.dispatch(update(null, () => '12.04.1961 20:15'));
   t.is(input.value, '04.04.1980 13:37');
 });
@@ -350,12 +332,12 @@ test('update with undefined ref', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   store.dispatch(update(undefined, () => '12.04.1961 20:15'));
   t.is(input.value, '04.04.1980 13:37');
 });
@@ -366,12 +348,12 @@ test('disable', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema} enabled={false}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   t.true(input.disabled);
 });
 
@@ -381,11 +363,11 @@ test('enabled by default', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   t.false(input.disabled);
 });
