+++ packages/vanilla/test/renderers/DateTimeField.test.tsx
@@ -1,23 +1,17 @@
-import '../../../test/helpers/setup';
+import { initJsonFormsStore } from '@jsonforms/test';
 import * as React from 'react';
 import test from 'ava';
 import {
   ControlElement,
   getData,
   HorizontalLayout,
-  initJsonFormsStore,
   JsonSchema,
   update
 } from '@jsonforms/core';
-import HorizontalLayoutRenderer from '../../src/layouts/horizontal.layout';
-import DateTimeField, { datetimeFieldTester } from '../../src/fields/datetime.field';
-import {
-  change,
-  findRenderedDOMElementWithTag,
-  renderIntoDocument,
-  scryRenderedDOMElementsWithTag
-} from '../../../test/helpers/binding';
+import HorizontalLayoutRenderer from '../../src/layouts/HorizontalLayout';
+import DateTimeField, { datetimeFieldTester } from '../../src/fields/DateTimeField';
 import { Provider } from 'react-redux';
+import * as TestUtils from 'react-dom/test-utils';
 
 test.beforeEach(t => {
   t.context.data = { 'foo': '1980-04-04T13:37:00.000Z' };
@@ -32,9 +26,7 @@ test.beforeEach(t => {
   };
   t.context.uischema = {
     type: 'Control',
-    scope: {
-      $ref: '#/properties/foo',
-    },
+    scope: '#/properties/foo',
   };
   t.context.styles = [
     {
@@ -57,18 +49,14 @@ test.failing('autofocus on first element', t => {
   };
   const firstControlElement: ControlElement = {
     type: 'Control',
-    scope: {
-      $ref: '#/properties/firstDate'
-    },
+    scope: '#/properties/firstDate',
     options: {
       focus: true
     }
   };
   const secondControlElement: ControlElement = {
     type: 'Control',
-    scope: {
-      $ref: '#/properties/secondDate'
-    },
+    scope: '#/properties/secondDate',
     options: {
       focus: true
     }
@@ -89,12 +77,12 @@ test.failing('autofocus on first element', t => {
     schema,
     uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <HorizontalLayoutRenderer schema={schema} uischema={uischema}/>
     </Provider>
   );
-  const inputs = scryRenderedDOMElementsWithTag(tree, 'input');
+  const inputs = TestUtils.scryRenderedDOMElementsWithTag(tree, 'input');
   t.not(document.activeElement, inputs[0]);
   t.is(document.activeElement, inputs[1]);
 });
@@ -102,9 +90,7 @@ test.failing('autofocus on first element', t => {
 test('autofocus active', t => {
   const uischema: ControlElement = {
     type: 'Control',
-    scope: {
-      $ref: '#/properties/foo'
-    },
+    scope: '#/properties/foo',
     options: {
       focus: true
     }
@@ -115,21 +101,19 @@ test('autofocus active', t => {
     uischema
   });
 
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   t.is(document.activeElement, input);
 });
 
 test('autofocus inactive', t => {
   const uischema: ControlElement = {
     type: 'Control',
-    scope: {
-      $ref: '#/properties/foo'
-    },
+    scope: '#/properties/foo',
     options: {
       focus: false
     }
@@ -139,33 +123,31 @@ test('autofocus inactive', t => {
     schema: t.context.schema,
     uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   t.false(input.autofocus);
 });
 
 test('autofocus inactive by default', t => {
   const uischema: ControlElement = {
     type: 'Control',
-    scope: {
-      $ref: '#/properties/foo'
-    }
+    scope: '#/properties/foo'
   };
   const store = initJsonFormsStore({
     data: t.context.data,
     schema: t.context.schema,
     uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   t.false(input.autofocus);
 });
 
@@ -234,13 +216,13 @@ test('render', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
 
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   t.is(input.type, 'datetime-local');
   t.is(input.value, '1980-04-04T13:37');
 });
@@ -251,14 +233,14 @@ test.cb('update via event', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   input.value = '1961-04-12T20:15';
-  change(input);
+  TestUtils.Simulate.change(input);
   setTimeout(
     () => {
       t.is(getData(store.getState()).foo, '1961-04-12T20:15:00.000Z');
@@ -274,12 +256,12 @@ test.cb('update via action', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   store.dispatch(update('foo', () => '1961-04-12T20:15:00.000Z'));
   setTimeout(
     () => {
@@ -296,12 +278,12 @@ test('update with null value', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   store.dispatch(update('foo', () => null));
   t.is(input.value, '');
 });
@@ -312,12 +294,12 @@ test('update with undefined value', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   store.dispatch(update('foo', () => undefined));
   t.is(input.value, '');
 });
@@ -328,12 +310,12 @@ test('update with wrong ref', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   store.dispatch(update('bar', () => 'Bar'));
   t.is(input.value, '1980-04-04T13:37');
 });
@@ -344,12 +326,12 @@ test('update with null ref', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   store.dispatch(update(null, () => '1961-04-12T20:15:00.000Z'));
   t.is(input.value, '1980-04-04T13:37');
 });
@@ -360,12 +342,12 @@ test('update with undefined ref', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   store.dispatch(update(undefined, () => '1961-04-12T20:15:00.000Z'));
   t.is(input.value, '1980-04-04T13:37');
 });
@@ -376,12 +358,12 @@ test('disable', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema} enabled={false}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   t.true(input.disabled);
 });
 
@@ -391,11 +373,11 @@ test('enabled by default', t => {
     schema: t.context.schema,
     uischema: t.context.uischema
   });
-  const tree = renderIntoDocument(
+  const tree = TestUtils.renderIntoDocument(
     <Provider store={store}>
       <DateTimeField schema={t.context.schema} uischema={t.context.uischema}/>
     </Provider>
   );
-  const input = findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
+  const input = TestUtils.findRenderedDOMElementWithTag(tree, 'input') as HTMLInputElement;
   t.false(input.disabled);
 });
